@@ -1,6 +1,6 @@
 package com.mk.spicyisland.koan.activities
 
-import android.support.v7.app.AppCompatActivity
+import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
 import com.mk.spicyisland.koan.R
 
