@@ -3,14 +3,14 @@ package com.mk.spicyisland.koan.fragments
 import android.content.Intent
 import android.net.Uri
 import android.os.Bundle
-import android.support.v4.app.Fragment
+import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import android.webkit.CookieManager
 import io.realm.Realm
 import kotlinx.android.synthetic.main.fragment_account.*
-import android.support.v7.app.AlertDialog
+import androidx.appcompat.app.AlertDialog
 import com.mk.spicyisland.koan.activities.FaqActivity
 import com.mk.spicyisland.koan.activities.LoginActivity
 import com.mk.spicyisland.koan.R
