@@ -1,9 +1,9 @@
 package com.mk.spicyisland.koan.fragments
 
 import android.content.Intent
-import android.databinding.DataBindingUtil
+import androidx.databinding.DataBindingUtil
 import android.os.Bundle
-import android.support.v4.app.Fragment
+import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
@@ -33,63 +33,63 @@ class BulletinBoardFragment : Fragment() {
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
-        setOnClickListeners()
+//        setOnClickListeners()
     }
 
-    private fun setOnClickListeners() {
-        lesson_bulletin_bulletin.setOnClickListener{
-            openWebView(0)
-        }
-
-        lesson_bulletin.setOnClickListener{
-            openWebView(1)
-        }
-
-        notification_bulletin.setOnClickListener{
-            openWebView(2)
-        }
-
-        individual_bulletin.setOnClickListener{
-            openWebView(3)
-        }
-
-        student_affairs_office_bulletin.setOnClickListener{
-            openWebView(4)
-        }
-
-        minor_course_bulletin.setOnClickListener{
-            openWebView(5)
-        }
-
-        teaching_profession_bulletin.setOnClickListener{
-            openWebView(6)
-        }
-
-        scholarship_bulletin.setOnClickListener{
-            openWebView(7)
-        }
-
-        career_bulletin.setOnClickListener{
-            openWebView(8)
-        }
-
-        school_life_bulletin.setOnClickListener{
-            openWebView(9)
-        }
-
-        study_abroad_student_bulletin.setOnClickListener{
-            openWebView(10)
-        }
-
-        study_abroad_bulletin.setOnClickListener{
-            openWebView(11)
-        }
-
-        other_bulletin.setOnClickListener{
-            openWebView(12)
-        }
-
-    }
+//    private fun setOnClickListeners() {
+//        lesson_bulletin_bulletin.setOnClickListener{
+//            openWebView(0)
+//        }
+//
+//        lesson_bulletin.setOnClickListener{
+//            openWebView(1)
+//        }
+//
+//        notification_bulletin.setOnClickListener{
+//            openWebView(2)
+//        }
+//
+//        individual_bulletin.setOnClickListener{
+//            openWebView(3)
+//        }
+//
+//        student_affairs_office_bulletin.setOnClickListener{
+//            openWebView(4)
+//        }
+//
+//        minor_course_bulletin.setOnClickListener{
+//            openWebView(5)
+//        }
+//
+//        teaching_profession_bulletin.setOnClickListener{
+//            openWebView(6)
+//        }
+//
+//        scholarship_bulletin.setOnClickListener{
+//            openWebView(7)
+//        }
+//
+//        career_bulletin.setOnClickListener{
+//            openWebView(8)
+//        }
+//
+//        school_life_bulletin.setOnClickListener{
+//            openWebView(9)
+//        }
+//
+//        study_abroad_student_bulletin.setOnClickListener{
+//            openWebView(10)
+//        }
+//
+//        study_abroad_bulletin.setOnClickListener{
+//            openWebView(11)
+//        }
+//
+//        other_bulletin.setOnClickListener{
+//            openWebView(12)
+//        }
+//
+//    }
 
     /**
      * TODO: ページの有効期限切れになるのでなんとかする
