@@ -1,7 +1,7 @@
 package com.mk.spicyisland.koan.models
 
-import android.arch.lifecycle.MutableLiveData
-import android.arch.lifecycle.ViewModel
+import androidx.lifecycle.MutableLiveData
+import androidx.lifecycle.ViewModel
 
 /**
  * LiveDataとして扱うためのクラス
